feat: generalize a serializer for Proto.
The example is still available under the test folder, under Settings
