+++ app/src/test/java/com/github/se/cyrcle/io/Settings.kt
@@ -1,5 +1,8 @@
 package com.github.se.cyrcle.io
 
+import java.io.InputStream
+import java.io.OutputStream
+
 data class Settings(
     val userName: String,
     val age: Int,
@@ -29,5 +32,25 @@ data class Settings(
     result = 31 * result + heightInCm.hashCode()
     result = 31 * result + isDeveloper.hashCode()
     return result
+  }
+
+  class Serializer :
+      ProtoGenericSerializer<Settings>(Settings("John Doe", 42, 420000, 4.2f, true)) {
+    override suspend fun readFrom(input: InputStream): Settings {
+      return Settings(
+          readString(input),
+          readInt32(input),
+          readLong64(input),
+          readFloat32(input),
+          readBoolean(input))
+    }
+
+    override suspend fun writeTo(t: Settings, output: OutputStream) {
+      writeString(output, t.userName)
+      writeInt32(output, t.age)
+      writeLong64(output, t.ageSinceEpoch)
+      writeFloat32(output, t.heightInCm)
+      writeBoolean(output, t.isDeveloper)
+    }
   }
 }
